#15 - Rename the folders of constants and helpers
