+++ practice-one/src/assets/icons/index.ts
@@ -4,4 +4,5 @@ export { default as Envelope } from './envelope-regular.svg';
 export { default as MagnifyingGlass } from './magnifying-glass-solid.svg';
 export { default as Pencil } from './pencil-solid.svg';
 export { default as Upload } from './upload-solid.svg';
+export { default as UserGroup } from './user-group-solid.svg';
 export { default as Xmark } from './xmark-solid.svg';
