+++ practice-one/src/components/DataDisplay/Icon/index.tsx
@@ -1,15 +1,19 @@
 type IconProps = {
-  src?: string;
+  src: string;
   width?: string;
   height?: string;
 };
 
 const Icon = ({
   src,
-  width = '20px',
-  height = '20px'
+  width = '16px',
+  height = '16px'
 }: IconProps) => {
-  return <img src={src} style={{ width: width, height: height }} />;
+  return <img
+    src={src}
+    width={width}
+    height={height} 
+  />;
 };
 
 export default Icon;
