+++ practice-one/src/components/Inputs/SearchBar/SearchBar.scss
@@ -7,35 +7,6 @@
   margin-bottom   : 20px;
   gap             : 20px;
 
-  &__magnifier--icon,
-  &__close--icon {
-    position: relative;
-
-    &:hover {
-      &::before {
-        @extend .hover-icon;
-      }
-    }
-
-    svg {
-      @extend .style-icon;
-
-      &:hover {
-        fill: $text-primary;
-      }
-    }
-  }
-
-  &__magnifier--icon {
-    margin-right: 10px;
-  }
-
-  &__close--icon {
-    svg {
-      font-size: $fs-medium;
-    }
-  }
-
   .hidden {
     display: none;
   }
