+++ practice-one/src/components/index.ts
@@ -1,4 +1,16 @@
-export { Avatar, Drawer, Icon, Status, Table } from './DataDisplay/index';
-export { Button, SearchBar, TextField } from './Inputs/index';
+export {
+  Avatar,
+  Drawer,
+  Icon,
+  Status,
+  Table
+} from './DataDisplay/index';
+
+export {
+  Button,
+  SearchBar,
+  TextField
+} from './Inputs/index';
+
 export { UserDetails } from './Surfaces/index';
