+++ practice-one/src/helpers/generate.ts
