+++ practice-one/src/styles/utilities/mixin.scss
@@ -23,24 +23,3 @@
     border: 1px solid $primary-color;
   }
 }
-
-.hover-icon {
-  content         : '';
-  position        : absolute; //Position the pseudo element
-  top             : 50%;
-  left            : 50%;
-  transform       : translate(-50%, -50%); //Center it horizontally and vertically.
-  border-radius   : 50%;
-  background-color: #f4f5f9;
-  width           : 40px;
-  height          : 40px;
-  cursor          : pointer;
-}
-
-.style-icon {
-  font-size: $fs-x-small;
-  fill     : #94a1b3;
-  position : relative;
-  z-index  : 1;
-  cursor   : pointer;
-}
