+++ practice-one/src/styles/utilities/variables.scss
@@ -1,5 +1,5 @@
 /* Common color begin*/
-$primary-color  : #1CA1C1;
+$primary-color  : #1ca1c1;
 $secondary-color: #fff;
 $tertiary-color : #313131;
 /* Common color end*/
